+++ apis/nucleus/src/components/listbox/assets/__tests__/get-list-sizes.test.jsx
@@ -42,21 +42,19 @@ describe('get-list-sizes', () => {
       columnCount: 4,
       columnWidth: 47.5,
       count: 200,
-      itemSize: 33,
+      itemSize: 29,
       listCount: 100,
       listHeight: 300,
+      maxCount: {
+        column: 706315,
+        row: 577000,
+      },
       overflowStyling: {
         overflowX: 'hidden',
       },
       rowCount: 25,
       scrollBarWidth: 10,
     });
-  });
-
-  it('should return expected itemSize with checkboxes true', () => {
-    args.checkboxes = true;
-    const sizes = getListSizes(args);
-    expect(sizes).toMatchObject({ itemSize: 40 });
   });
 
   it('dense should override itemSize', () => {
@@ -77,6 +75,10 @@ describe('get-list-sizes', () => {
       itemSize: 45,
       listCount: 100,
       listHeight: 300,
+      maxCount: {
+        column: 706315,
+        row: 577000,
+      },
       overflowStyling: {
         overflowY: 'hidden',
       },
@@ -99,5 +101,58 @@ describe('get-list-sizes', () => {
     args.layout.layoutOptions.layoutOrder = 'column';
     const sizes = getListSizes(args);
     expect(sizes.itemSize).toEqual(45); // itemSize + padding = 33 + 12 = 45
+  });
+
+  it('maxRowCount should limit listCount and rowCount, in column layout', () => {
+    args.layoutOrder = 'column';
+    const maxRowCount = 577000;
+    const columnCount = 4;
+    args.listCount = maxRowCount * columnCount + 1;
+    const sizes = getListSizes(args);
+    expect(sizes).toEqual({
+      columnCount: 4,
+      columnWidth: 47.5,
+      count: 200,
+      itemSize: 29,
+      listCount: 2308000,
+      listHeight: 300,
+      maxCount: {
+        column: 706315,
+        row: 577000,
+      },
+      overflowStyling: {
+        overflowX: 'hidden',
+      },
+      rowCount: 577000,
+      scrollBarWidth: 10,
+    });
+  });
+
+  it('maxColumnCount should limit listCount and columnCount, in grid layout', () => {
+    args.layout.layoutOptions.dataLayout = 'grid';
+    args.height = 100;
+    args.layout.layoutOptions.layoutOrder = 'column';
+    const rowCount = 3;
+    const columnCount = 493382;
+    args.listCount = rowCount * columnCount + 1;
+    const limitedListCount = args.listCount - 1;
+    const sizes = getListSizes(args);
+    expect(sizes).toEqual({
+      columnCount,
+      columnWidth: 68,
+      count: 200,
+      itemSize: 29,
+      listCount: limitedListCount,
+      listHeight: 100,
+      maxCount: {
+        column: columnCount,
+        row: 577000,
+      },
+      overflowStyling: {
+        overflowY: 'hidden',
+      },
+      rowCount,
+      scrollBarWidth: 10,
+    });
   });
 });