+++ apis/nucleus/src/components/listbox/components/grid-list-components/grid-list-components.jsx
@@ -4,6 +4,7 @@ import RowColumn from '../ListBoxRowColumn';
 import getFrequencyAllowed from './frequency-allowed';
 import deriveRenderOptions from './derive-render-options';
 import getStyledComponents, { classes } from './styled-components';
+import handleSetOverflowDisclaimer from './set-overflow-disclaimer';
 
 const { StyledFixedSizeList, StyledFixedSizeGrid } = getStyledComponents();
 
@@ -29,6 +30,9 @@ export default function getListBoxComponents({
   listLayout,
   sizes,
   listCount,
+  overflowDisclaimer,
+  setScrollPosition,
+  focusListItems,
 }) {
   const { layoutOptions = {}, frequencyMax } = layout || {};
   const { listHeight, itemSize, rowCount, columnCount } = sizes || {};
@@ -58,11 +62,14 @@ export default function getListBoxComponents({
       select,
       confirm: () => selections?.confirm.call(selections),
       cancel: () => selections?.cancel.call(selections),
+      setScrollPosition,
     },
     frequencyMax,
     histogram,
     keyboard,
     showGray,
+    dataOffset: local.current.dataOffset,
+    focusListItems,
   };
 
   const List = ({ onItemsRendered, ref }) => {
@@ -77,12 +84,22 @@ export default function getListBoxComponents({
         width={width}
         itemCount={listCount}
         layout={listLayout}
-        itemData={{ ...commonItemData }}
+        itemData={{ ...commonItemData, listCount }}
         itemSize={itemSize}
         onItemsRendered={(renderProps) => {
           if (scrollState) {
             scrollState.setScrollPos(renderProps.visibleStopIndex);
           }
+          handleSetOverflowDisclaimer({
+            renderProps,
+            layoutOptions,
+            maxCount: sizes.maxCount,
+            columnCount,
+            rowCount,
+            overflowDisclaimer,
+            qCardinal: layout?.qListObject?.qDimensionInfo?.qCardinal,
+            dataOffset: local.current.dataOffset,
+          });
           onItemsRendered({ ...renderProps });
         }}
         ref={ref}
@@ -106,6 +123,16 @@ export default function getListBoxComponents({
         layoutOrder,
         rowCount,
         columnCount,
+      });
+      handleSetOverflowDisclaimer({
+        renderProps,
+        layoutOptions,
+        maxCount: sizes.maxCount,
+        columnCount,
+        rowCount,
+        overflowDisclaimer,
+        qCardinal: layout?.qListObject?.qDimensionInfo?.qCardinal,
+        dataOffset: local.current.dataOffset,
       });
       onItemsRendered(renderOptions);
     };