+++ apis/nucleus/src/components/listbox/hooks/__tests__/useItemsLoader.test.jsx
@@ -19,6 +19,7 @@ describe('useItemsLoader', () => {
       local: {
         current: {
           queue: [],
+          dataOffset: 0,
         },
       },
       loaderRef: { current: undefined },