+++ commands/sense/lib/build.js
@@ -4,8 +4,8 @@ const fs = require('fs-extra');
 const rollup = require('rollup');
 const { nodeResolve } = require('@rollup/plugin-node-resolve');
 const common = require('@rollup/plugin-commonjs');
-const babel = require('rollup-plugin-babel');
-const { terser } = require('rollup-plugin-terser');
+const babel = require('@rollup/plugin-babel');
+const terser = require('@rollup/plugin-terser');
 
 async function build(argv) {
   const cwd = process.cwd();
@@ -93,6 +93,7 @@ async function build(argv) {
         nodeResolve(),
         common(),
         babel({
+          babelHelpers: 'bundled',
           babelrc: false,
           exclude: [/node_modules/],
           presets: [