+++ apis/nucleus/CHANGELOG.md
@@ -3,6 +3,53 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.3](https://github.com/qlik-oss/nebula.js/compare/v3.1.2...v3.1.3) (2023-02-09)
+
+### Features
+
+- expose close event for listbox popover ([#1070](https://github.com/qlik-oss/nebula.js/issues/1070)) ([e92aeb0](https://github.com/qlik-oss/nebula.js/commit/e92aeb058a38758eb2a7f957c69fd065fca7da3e))
+- **Listbox:** load end of data ([#1068](https://github.com/qlik-oss/nebula.js/issues/1068)) ([ae14bc4](https://github.com/qlik-oss/nebula.js/commit/ae14bc48560164f9f936ff0038e69dc5ab84df81))
+- wildcard search ([#1075](https://github.com/qlik-oss/nebula.js/issues/1075)) ([a2ba9c2](https://github.com/qlik-oss/nebula.js/commit/a2ba9c2a3013be78c73e155fe014f2ad7d93eb7b))
+
+## [3.1.2](https://github.com/qlik-oss/nebula.js/compare/v3.1.1...v3.1.2) (2023-01-30)
+
+### Features
+
+- created object now can extend initial properties ([#1063](https://github.com/qlik-oss/nebula.js/issues/1063)) ([c14a1e6](https://github.com/qlik-oss/nebula.js/commit/c14a1e64a4a03d3aaef72949f01ec358d8ce2a22))
+- listbox popover api ([#1067](https://github.com/qlik-oss/nebula.js/issues/1067)) ([744fe21](https://github.com/qlik-oss/nebula.js/commit/744fe2189d3e0ade89b2855980c9e7d429765053))
+- **Listbox:** overflow disclaimer ([#1059](https://github.com/qlik-oss/nebula.js/issues/1059)) ([ed97b08](https://github.com/qlik-oss/nebula.js/commit/ed97b081729053262fa0b69c23509a5d9f3f63cd))
+
+### Reverts
+
+- Revert "fix!: created object now extend initial properties" ([6e07403](https://github.com/qlik-oss/nebula.js/commit/6e0740369ceaf9deefd15e5e560bbff499a1fcda))
+
+## [3.1.1](https://github.com/qlik-oss/nebula.js/compare/v3.1.0...v3.1.1) (2023-01-22)
+
+**Note:** Version bump only for package @nebula.js/nucleus
+
+# [3.1.0](https://github.com/qlik-oss/nebula.js/compare/v3.1.0-alpha.5...v3.1.0) (2023-01-17)
+
+**Note:** Version bump only for package @nebula.js/nucleus
+
+# [3.1.0-alpha.5](https://github.com/qlik-oss/nebula.js/compare/v3.1.0-alpha.4...v3.1.0-alpha.5) (2023-01-13)
+
+**Note:** Version bump only for package @nebula.js/nucleus
+
+# [3.1.0-alpha.4](https://github.com/qlik-oss/nebula.js/compare/v3.1.0-alpha.3...v3.1.0-alpha.4) (2023-01-13)
+
+### Bug Fixes
+
+- **listbox:** Listobx's selection dropdown anchor issue ([#1048](https://github.com/qlik-oss/nebula.js/issues/1048)) ([270769c](https://github.com/qlik-oss/nebula.js/commit/270769c7a559c5c8cd6c169eb81d726e6fc817cd))
+- prevent rendering empty values ([#1035](https://github.com/qlik-oss/nebula.js/issues/1035)) ([02e1ec3](https://github.com/qlik-oss/nebula.js/commit/02e1ec372d2011a61525478795b3d93d6c6a691f))
+
+### Features
+
+- add checkboxes, histogram as properties ([#1028](https://github.com/qlik-oss/nebula.js/issues/1028)) ([a333cf9](https://github.com/qlik-oss/nebula.js/commit/a333cf9dd175844070acbe89dd23c6b8093565e1))
+- add disclaimer listbox search ([#1030](https://github.com/qlik-oss/nebula.js/issues/1030)) ([4a5b407](https://github.com/qlik-oss/nebula.js/commit/4a5b40723271def71ba42134a76a37f002012100))
+- listbox grid layout ([#994](https://github.com/qlik-oss/nebula.js/issues/994)) ([897275a](https://github.com/qlik-oss/nebula.js/commit/897275ae7c79b8214af9c99ce4e9288bbf288690))
+- listbox text alignment ([#933](https://github.com/qlik-oss/nebula.js/issues/933)) ([4fec8b0](https://github.com/qlik-oss/nebula.js/commit/4fec8b06d35f86c391d6eb972720a9354a44a7ca))
+- **Listbox:** row height ([#1029](https://github.com/qlik-oss/nebula.js/issues/1029)) ([1076ed4](https://github.com/qlik-oss/nebula.js/commit/1076ed4bb5e2a64e9c1592695d99882d42215401))
+
 # [3.1.0-alpha.3](https://github.com/qlik-oss/nebula.js/compare/v3.1.0-alpha.2...v3.1.0-alpha.3) (2022-12-12)
 
 ### Bug Fixes